Separate UI logic (MainActivity) from Business logic (CipherViewModel & Utils)
